Add Create/Optimize table command base classes and refactor delta-33x
- Add GpuCreateDeltaTableCommandBase in common directory
- Add OptimizeTableCommandMetaBase and GpuOptimizeTableCommand in common directory
- Refactor delta-33x GpuCreateDeltaTableCommand and OptimizeTableCommandMeta to use base classes

Tested:
- Delta 3.3.0 (buildver=356): BUILD SUCCESS
- Delta 4.0.0 (buildver=400): BUILD SUCCESS
